+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/Checkbox/CheckboxConfiguration.swift
@@ -82,7 +82,7 @@ struct CheckboxConfiguration: View {
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(model.isError)
 
-            Toggle("app_components_common_onError_label", isOn: $model.isError)
+            Toggle("app_components_common_error_label", isOn: $model.isError)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(!model.enabled)