+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/Checkbox/CheckboxIndeterminateElement.swift
@@ -20,7 +20,7 @@ struct CheckboxIndeterminateElement: DesignToolboxElement {
     let pageDescription: AnyView
 
     init() {
-        name = "app_components_checkbox_threeStates_indicatorOnly_label".localized()
+        name = "app_components_checkbox_indeterminateCheckbox_label".localized()
         image = Image(decorative: "il_component_checkbox").renderingMode(.original)
         pageDescription = AnyView(DesignToolboxElementPage(
             name: name,