+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/CheckboxItem/CheckboxItemPage.swift
@@ -74,21 +74,17 @@ private struct CheckboxItemDemo: View {
 
     var body: some View {
         OUDSCheckboxItem(isOn: $model.indicatorState,
-                         label: model.labelTextContent,
-                         helper: helperTextContent,
+                         label: model.labelText,
+                         helper: model.helperText,
                          icon: icon,
                          flipIcon: model.flipIcon,
-                         isReversed: model.layoutOrientation == .reversed,
+                         isReversed: model.isReversed,
                          isError: model.isError,
                          isReadOnly: model.isReadOnly,
                          hasDivider: model.divider)
         .disabled(!model.enabled)
         .padding(.all, theme.spaces.spaceFixedMedium)
         .designToolboxColoredSurface(false)
-    }
-
-    private var helperTextContent: String? {
-        model.helperText ? model.helperTextContent : nil
     }
 
     // Need here that system name, a11y managed in component