+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/CheckboxItem/CheckboxlItemConfiguration.swift
@@ -29,10 +29,6 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
-    @Published var helperText: Bool {
-        didSet { updateCode() }
-    }
-
     @Published var icon: Bool {
         didSet { updateCode() }
     }
@@ -49,15 +45,15 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
-    @Published var layoutOrientation: DesignToolboxLayoutOrientation {
+    @Published var isReverted: Bool {
         didSet { updateCode() }
     }
 
-    @Published var labelTextContent: String {
+    @Published var labelText: String {
         didSet { updateCode() }
     }
 
-    @Published var helperTextContent: String {
+    @Published var helperText: String {
         didSet { updateCode() }
     }
 
@@ -68,12 +64,11 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
         isError = false
         isReadOnly = false
         enabled = true
-        helperText = true
         icon = true
-        layoutOrientation = .default
+        isReverted = false
         divider = false
-        labelTextContent = String(localized: "app_components_checkbox_label_text")
-        helperTextContent = String(localized: "app_components_checkbox_helperText_text")
+        labelText = String(localized: "app_components_common_label_label")
+        helperText = String(localized: "app_components_controlItem_helperText_label")
     }
 
     deinit { }
@@ -84,7 +79,7 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
     override func updateCode() {
         code =
           """
-        OUDSCheckboxItem(selection: $selection, labelText: \"\(labelTextContent)\"\(helperTextPatern)\(iconPatern)\(isReversedPattern)\(isErrorPattern)\(isReadOnlyPattern)\(dividerPatern))
+        OUDSCheckboxItem(selection: $selection, labelText: \"\(labelText)\"\(helperTextPatern)\(iconPatern)\(isReversedPattern)\(isErrorPattern)\(isReadOnlyPattern)\(dividerPatern))
         \(disableCode)
         """
     }
@@ -95,7 +90,7 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
     }
 
     private var helperTextPatern: String {
-        helperText ? ", helperText: \"\(helperTextContent)\")" : ""
+        helperText.isEmpty ? "" : ", helper: \"\(helperText)\""
     }
 
     private var iconPatern: String {
@@ -103,7 +98,7 @@ final class CheckboxItemConfigurationModel: ComponentConfiguration {
     }
 
     private var isReversedPattern: String {
-        layoutOrientation == .reversed ? ", isReversed: true" : ""
+        ", isReversed: \(isReversed)"
     }
 
     private var isErrorPattern: String {
@@ -142,39 +137,35 @@ struct CheckboxItemConfiguration: View {
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(model.isError || model.isReadOnly)
 
-            Toggle("app_components_common_onError_label", isOn: $model.isError)
+            Toggle("app_components_common_error_label", isOn: $model.isError)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(!model.enabled || model.isReadOnly)
 
-            Toggle("app_components_common_readOnly_label", isOn: $model.isReadOnly)
+            Toggle("app_components_controlItem_readOnly_label", isOn: $model.isReadOnly)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(!model.enabled || model.isError)
 
-            Toggle("app_components_checkbox_helperText_label", isOn: $model.helperText)
+            Toggle("app_components_controlItem_helperText_label", isOn: $model.helperText)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
 
-            Toggle("app_components_common_icon_label", isOn: $model.icon)
+            Toggle("app_components_controlItem_icon_label", isOn: $model.icon)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
 
-            Toggle("app_components_common_divider_label", isOn: $model.divider)
+            Toggle("app_components_controlItem_divider_label", isOn: $model.divider)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
 
-            DesignToolboxChoicePicker(title: "app_components_common_orientation_label", selection: $model.layoutOrientation) {
-                ForEach(DesignToolboxLayoutOrientation.allCases, id: \.id) { orientation in
-                    Text(LocalizedStringKey(orientation.description)).tag(orientation)
-                }
-            }
+            Toggle("app_components_controlItem_reversed_label", isOn: $model.isReversed)
+                .typeHeadingMedium(theme)
+                .oudsForegroundStyle(theme.colors.colorContentDefault)
 
-            DisclosureGroup("app_components_common_editContent_label") {
-                DesignToolboxTextField(text: $model.labelTextContent, prompt: "app_components_common_userText_prompt", title: "app_components_common_labelText_label")
-                if model.helperText {
-                    DesignToolboxTextField(text: $model.helperTextContent, prompt: "app_components_common_userText_prompt", title: "app_components_common_helperText_label")
-                }
+            DesignToolboxContentModiferDisclosure {
+                DesignToolboxTextField(text: $model.labelText)
+                DesignToolboxTextField(text: $model.helperText)
             }
         }
     }