+++ DesignToolbox/DesignToolbox/Pages/Components/Switch/Switchitem/SwitchItemConfiguration.swift
@@ -29,7 +29,7 @@ final class SwitchItemConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
-    @Published var helperText: Bool {
+    @Published var helperText: String {
         didSet { updateCode() }
     }
 
@@ -53,15 +53,11 @@ final class SwitchItemConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
-    @Published var layoutOrientation: DesignToolboxLayoutOrientation {
+    @Published var isReversed: Bool {
         didSet { updateCode() }
     }
 
-    @Published var labelTextContent: String {
-        didSet { updateCode() }
-    }
-
-    @Published var helperTextContent: String {
+    @Published var labelText: String {
         didSet { updateCode() }
     }
 
@@ -72,13 +68,12 @@ final class SwitchItemConfigurationModel: ComponentConfiguration {
         isError = false
         isReadOnly = false
         enabled = true
-        helperText = true
+        helperText = String(localized: "app_components_controlItem_helperText_label")
         icon = true
         flipIcon = false
-        layoutOrientation = .default
+        isReversed = false
         divider = true
-        labelTextContent = String(localized: "app_components_switch_label_text")
-        helperTextContent = String(localized: "app_components_switch_helperText_text")
+        labelText = String(localized: "app_components_common_label_label")
     }
 
     deinit { }
@@ -89,7 +84,7 @@ final class SwitchItemConfigurationModel: ComponentConfiguration {
     override func updateCode() {
         code =
           """
-        OUDSSwitchItem(isOn: $isOn, label: "\(labelTextContent)"\(helperTextPatern)\(iconPatern)\(flipIconPattern)\(isReversedPattern)\(isErrorPattern)\(isReadOnlyPattern)\(dividerPatern))
+        OUDSSwitchItem(isOn: $isOn, label: "\(labelText)"\(helperTextPatern)\(iconPatern)\(flipIconPattern)\(isReversedPattern)\(isErrorPattern)\(isReadOnlyPattern)\(dividerPatern))
         \(disableCode)
         """
     }
@@ -100,7 +95,7 @@ final class SwitchItemConfigurationModel: ComponentConfiguration {
     }
 
     private var helperTextPatern: String {
-        helperText ? ", helper: \"\(helperTextContent)\"" : ""
+        helperText.isEmpty ? "" : ", helper: \"\(helperText)\""
     }
 
     private var iconPatern: String {
@@ -112,7 +107,7 @@ final class SwitchItemConfigurationModel: ComponentConfiguration {
     }
 
     private var isReversedPattern: String {
-        layoutOrientation == .reversed ? ", isReversed: true" : ""
+        ", isReversed: \(isReversed)"
     }
 
     private var isErrorPattern: String {
@@ -143,57 +138,41 @@ struct SwitchItemConfiguration: View {
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(model.isError || model.isReadOnly)
 
+            Toggle("app_components_controlItem_icon_label", isOn: $model.icon)
+                .typeHeadingMedium(theme)
+                .oudsForegroundStyle(theme.colors.colorContentDefault)
+
+            Toggle("app_components_controlItem_flipIcon_label", isOn: $model.flipIcon)
+                .typeHeadingMedium(theme)
+                .oudsForegroundStyle(theme.colors.colorContentDefault)
+                .disabled(!model.icon)
+
+            Toggle("app_components_controlItem_divider_label", isOn: $model.divider)
+                .typeHeadingMedium(theme)
+                .oudsForegroundStyle(theme.colors.colorContentDefault)
+
+            Toggle("app_components_controlItem_reversed_label", isOn: $model.isReversed)
+                .typeHeadingMedium(theme)
+                .oudsForegroundStyle(theme.colors.colorContentDefault)
+
             Toggle("app_common_enabled_label", isOn: $model.enabled)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(model.isError || model.isReadOnly)
 
-            Toggle("app_components_common_onError_label", isOn: $model.isError)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-                .disabled(!model.enabled || model.isReadOnly)
-
-            Toggle("app_components_common_readOnly_label", isOn: $model.isReadOnly)
+            Toggle("app_components_controlItem_readOnly_label", isOn: $model.isReadOnly)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(!model.enabled || model.isError)
 
-            Toggle("app_components_common_helperText_label", isOn: $model.helperText)
+            Toggle("app_components_common_error_label", isOn: $model.isError)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
+                .disabled(!model.enabled || model.isReadOnly)
 
-            Toggle("app_components_common_icon_label", isOn: $model.icon)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-
-            if model.icon {
-                Toggle("app_components_common_flipIcon_label", isOn: $model.flipIcon)
-                    .typeHeadingMedium(theme)
-                    .oudsForegroundStyle(theme.colors.colorContentDefault)
-            }
-
-            Toggle("app_components_common_divider_label", isOn: $model.divider)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-
-            DesignToolboxChoicePicker(title: "app_components_common_orientation_label", selection: $model.layoutOrientation) {
-                ForEach(DesignToolboxLayoutOrientation.allCases, id: \.id) { orientation in
-                    Text(LocalizedStringKey(orientation.description)).tag(orientation)
-                }
-            }
-
-            DisclosureGroup("app_components_common_editContent_label") {
-                DesignToolboxTextField(
-                    text: $model.labelTextContent,
-                    prompt: "app_components_common_userText_prompt",
-                    title: "app_components_common_labelText_label")
-
-                if model.helperText {
-                    DesignToolboxTextField(
-                        text: $model.helperTextContent,
-                        prompt: "app_components_common_userText_prompt",
-                        title: "app_components_common_helperText_label")
-                }
+            DesignToolboxEditContentDisclosure {
+                DesignToolboxTextField(text: $model.labelText)
+                DesignToolboxTextField(text: $model.helperText)
             }
         }
     }