+++ DesignToolbox/DesignToolbox/Pages/Components/Switch/Switchitem/SwitchItemPage.swift
@@ -74,21 +74,17 @@ private struct SwitchItemDemo: View {
 
     var body: some View {
         OUDSSwitchItem(isOn: $model.selection,
-                       label: model.labelTextContent,
-                       helper: helperTextContent,
+                       label: model.labelText,
+                       helper: model.helperText,
                        icon: icon,
                        flipIcon: model.flipIcon,
-                       isReversed: model.layoutOrientation == .reversed,
+                       isReversed: model.isReversed,
                        isError: model.isError,
                        isReadOnly: model.isReadOnly,
                        hasDivider: model.divider)
         .disabled(!model.enabled)
         .padding(.all, theme.spaces.spaceFixedMedium)
         .designToolboxColoredSurface(false)
-    }
-
-    private var helperTextContent: String? {
-        model.helperText ? model.helperTextContent : nil
     }
 
     // Need here that system name, a11y managed in component