+++ DesignToolbox/DesignToolbox/Pages/Tokens/Dimension/Space/SpaceTokenCommonIllustration.swift
@@ -85,6 +85,7 @@ struct SpaceTokenVariant<TokenIllustration>: View where TokenIllustration: View 
 struct SpaceCommonIllustration: View {
 
     @Environment(\.theme) private var theme
+    @Environment(\.colorScheme) private var colorScheme
 
     enum Padding {
         case top(SpaceIllustrationIcon.Asset?)
@@ -110,10 +111,8 @@ struct SpaceCommonIllustration: View {
 
             switch padding {
             case .topLeading:  // ZStack topleading
-                DesignToolboxTokenIllustrationBackground()
-                    .padding(.top, dimension)
-                    .padding(.leading, dimension)
-                    .oudsBackground(theme.colors.colorContentStatusInfo)
+                SpaceIllustrationRectangle(width: dimension)
+                SpaceIllustrationRectangle(height: dimension)
             case .leading(let asset): // ZStack alignment leading
                 HStack(alignment: .center, spacing: theme.spaces.spaceFixedNone) {
                     SpaceIllustrationRectangle(width: dimension)