+++ CHANGELOG.md
@@ -11,14 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Library] Switch component ([#405](https://github.com/Orange-OpenSource/ouds-ios/issues/405))
 - [Tool] New system to generate, build ad upload documentation ([#628](https://github.com/Orange-OpenSource/ouds-ios/issues/628))
 - [Library] Add divider component ([#511](https://github.com/Orange-OpenSource/ouds-ios/issues/511))
-- [Library] Add divider component ([#511](https://github.com/Orange-OpenSource/ouds-ios/issues/511))
-- [Library] Let control-item-based component flip the icons programatically or not ([#556](https://github.com/Orange-OpenSource/ouds-ios/issues/556))
 - [Library] Radio picker component ([#586](https://github.com/Orange-OpenSource/ouds-ios/issues/586))
 - [Library] Let control-item-based component flip the icons programatically or not ([#556](https://github.com/Orange-OpenSource/ouds-ios/issues/556))
 
 ### Changed
 
+- [DemoApp] Update wording keys ([#623](https://github.com/Orange-OpenSource/ouds-ios/issues/623))
 - [DesignToolbox] Udpate component illustrations ([#618](https://github.com/Orange-OpenSource/ouds-ios/issues/618))
+- [DesignToolbox] Update configuration section to group content edition in a same group ([#637](https://github.com/Orange-OpenSource/ouds-ios/issues/637))
 - [DesignToolbox] Udpate language translations ([#622](https://github.com/Orange-OpenSource/ouds-ios/issues/622))
 - [Tool] Fix Gitleaks configuration file and update documentation about version in use
 - [Library] Renaming of API ([#608](https://github.com/Orange-OpenSource/ouds-ios/issues/608))